BACKEND/backend/backenddemo/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.11
DateTime==4.3
Django==4.0.2
django-cors-headers==3.11.0
djangorestframework==3.13.1
idna==3.3
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.8
zope.interface==5.4.0
